# emas/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.ChatbotApproval:
    properties:
      args_json:
        type: string
      conversation_id:
        type: string
      created_at:
        type: string
      decided_at:
        type: string
      decided_by:
        type: string
      execution_error:
        type: string
      id:
        type: string
      idempotency_key:
        type: string
      method:
        type: string
      path:
        type: string
      request_id:
        type: string
      requested_by:
        type: string
      result_snapshot_id:
        type: string
      risk_summary:
        type: string
      side_effect_level:
        type: string
      status:
        description: PENDING, APPROVED, REJECTED, EXECUTED, EXPIRED, FAILED
        type: string
      tool_name:
        type: string
      turn_audit_id:
        type: string
    type: object
...
  emas_internal_handler_dto.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Factory operations API.
  title: eMas Factory API
  version: "1.0"
paths:
...
  /chatbot/approval/{id}:
    get:
      consumes:
      - application/json
      description: Get an approval by ID
      parameters:
      - description: Approval ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/emas_internal_handler_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ChatbotApproval'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Get an approval by ID
      tags:
      - chatbot approval
  /chatbot/approval/{id}/approve:
    post:
      consumes:
      - application/json
      description: Approve an approval
      parameters:
      - description: Approval ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/emas_internal_handler_dto.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Approve an approval
      tags:
      - chatbot approval
  /chatbot/approval/{id}/reject:
    post:
      consumes:
      - application/json
      description: Reject an approval
      parameters:
      - description: Approval ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/emas_internal_handler_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ChatbotApproval'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Reject an approval
      tags:
      - chatbot approval
  /chatbot/approval/pending:
    get:
      consumes:
      - application/json
      description: List pending approvals
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/emas_internal_handler_dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.ChatbotApproval'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: List pending approvals
      tags:
      - chatbot approval
...
swagger: "2.0"